Fixed these three errors
* `embed.fields.2.value: This field is required`

It looks like embeds fail when the value is `""`. If the field has a value, only then add it to the embed; otherwise skip adding that value.

* `In content: Must be 2000 or fewer in length.` in response to a bad command

Splitting up the help into commands examples and sending these messages sequentially fixed this issue.

* `ValueError: No closing quotation` - arguments need to be better sanitized

Added check for odd numbers of single or double quotes. It's silly that shlex doesn't catch and remediate this.