Fixed issue with game name mismatch
BGA has certain names for games in game creation that do not match what
they are called when the game is ongoing. For example, the name of a
table is sechsnimmt when the game is 6nimmt.

This commit adds an alias function to translate between the expected
name in game creation and the actual one used in the table json.